Fix a bug in the implementation of pipe
The shell needs to handle 3 cases:

1. The main process should forks a child A, which will create the fd
and fork another child B. A will execute the first command, and B will
continue to parse the rest of the line.

2. If in a forked child C due to pipe, and there is another pipe, C
should fork only once (child D). C will execute the command and D will
continue parsing.

3. If in a forked child E due to pipe and there is no more pipe, E just
execute the command.

The bug is the result of not distinguishing Case 1 and 2 so that in
Case 2 two processes will be created instead of one.